feat(api): Implement login functionality
- Fix environment variable issue.

[Finishes 161462334]
